+++ src/full/Agda/Compiler/UHC/Pragmas/Parse.hs
@@ -11,6 +11,7 @@ module Agda.Compiler.UHC.Pragmas.Parse
   , coreExprToCExpr
   , parseCoreData
   , parseCoreConstrs
+  , parseUHCImport
   )
 
 where
@@ -116,3 +117,10 @@ isMagicEntity _ _ _ | otherwise = return Nothing
 -- A syntactally correct magic name is NOT necessarily a valid magic name.
 isMagic :: String -> Bool
 isMagic xs = "__" `isPrefixOf` xs && "__" `isSuffixOf` xs
+
+-- | Parse a IMPORT_UHC expression
+parseUHCImport :: MonadTCM m => ForeignCode -> m String
+parseUHCImport (ForeignCode r s) = liftTCM $ setCurrentRange r $
+  case words s of
+    "__IMPORT__" : ws -> return $ unwords ws
+    _                 -> typeError $ GenericError $ "FOREIGN UHC code must be __IMPORT__ <module>"