+++ src/full/Agda/TypeChecking/Reduce.hs
@@ -1260,14 +1260,12 @@ instance InstantiateFull Clause where
 
 instance InstantiateFull Interface where
     instantiateFull' (Interface h ms mod scope inside
-                               sig display b hsImports hsImportsUHC hsCode highlighting pragmas patsyns) =
+                               sig display b foreignCode highlighting pragmas patsyns) =
         Interface h ms mod scope inside
             <$> instantiateFull' sig
             <*> instantiateFull' display
             <*> instantiateFull' b
-            <*> return hsImports
-            <*> return hsImportsUHC
-            <*> return hsCode
+            <*> return foreignCode
             <*> return highlighting
             <*> return pragmas
             <*> return patsyns